requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
CounterFit==0.1.4.dev9
counterfit-connection==0.1.0.dev5
counterfit-shims-grove==0.1.4.dev5
counterfit-shims-picamera==0.1.0.dev5
counterfit-shims-serial==0.1.0.dev3
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dnspython==2.7.0
eventlet==0.33.3
filetype==1.2.0
Flask==2.1.2
Flask-SocketIO==5.2.0
fonttools==4.55.8
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.7
inference-sdk==0.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
msal==1.31.1
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
paho-mqtt==2.1.0
pillow==10.4.0
propcache==0.2.1
py-cpuinfo==9.0.0
pycparser==2.22
PyJWT==2.10.1
pynmea2==1.19.0
pyodbc==5.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.53
scipy==1.15.1
simple-websocket==1.1.0
six==1.17.0
supervision==0.25.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.18.3
